Split the shared library into nested routes and components
Fix bug where the shared flow page was not loaded after refresh
