+++ frontend/src/pages/team/Library/Blueprints.vue
@@ -1,0 +1,13 @@
+<template>
+    <h1>Blueprints</h1>
+</template>
+
+<script>
+export default {
+    name: 'BluePrints'
+}
+</script>
+
+<style scoped lang="scss">
+
+</style>
