+++ frontend/src/pages/team/Library/routes.js
@@ -1,1 +1,7 @@
-export default []
+import Blueprints from './Blueprints.vue'
+import TeamLibrary from './TeamLibrary.vue'
+
+export default [
+    { name: 'TeamSharedLibraryBlueprints', path: 'blueprints', component: Blueprints },
+    { name: 'TeamSharedLibraryTeamLibrary', path: 'team-library/:entryPath*', component: TeamLibrary }
+]
